requirements.txt
aiosqlite==0.17.0
python-dotenv==0.17.0
discord.py==1.7.1